Fixed annotations

+++ error_contribution.py
@@ -1,4 +1,5 @@
 import argparse
+import os
 
 from models.gat import GATNet
 from models.gat_gcn import GAT_GCN
@@ -22,22 +23,50 @@ import pandas as pd
 import numpy as np
 import matplotlib.pyplot as plt
 import json
+from collections import OrderedDict
+from rdkit import Chem
+
+def find_matching_keys(file1_path, file2_path, canonicalize=False):
+    """
+    Finds keys in the second JSON file that have identical values to those in the first JSON file.
+    
+    Parameters:
+        file1_path (str): Path to the first JSON file.
+        file2_path (str): Path to the second JSON file.
+    
+    Returns:
+        dict: A dictionary where keys from the first JSON file map to lists of keys from the second JSON file
+              that have the same values.
+    """
+    # Load JSON files
+    with open(file1_path, 'r') as f1:
+        data1 = json.load(f1)
+
+    data2 = json.load(open(file2_path), object_pairs_hook=OrderedDict) 
+
+    if canonicalize==True:
+        data2 = {key: Chem.MolToSmiles(Chem.MolFromSmiles(value)) for key, value in data2.items()}
+
+    # Create a mapping of values to keys for the second JSON file
+    value_to_keys = {}
+    for key, value in data2.items():
+        if value not in value_to_keys:
+            value_to_keys[value] = []
+        value_to_keys[value].append(key)
+
+    # Create a result dictionary where keys from the second JSON file map to lists of values from the first JSON file
+    result = {}
+    for key, value in data2.items():
+        if value in data1:
+            if key not in result:
+                result[key] = []
+            result[key].append(value)
+
+    return result
 
 def calculate_mae(group):
     mae = (group['affinity'] - group['prediction']).abs().median()
     return mae
-
-# Function to plot MAE
-def plot_mae(df, x_col, y_col, title, xlabel, ylabel, highlight_points, ax):
-    # sns.scatterplot(data=df, x=x_col, y=y_col, ax=ax, s=10)
-    ax.scatter(df[x_col], df[y_col], s=10)
-    for point in highlight_points:
-        x = df[df[x_col] == point][x_col].values[0]
-        y = df[df[x_col] == point][y_col].values[0]
-        ax.annotate(point, (x, y), textcoords="offset points", xytext=(0,10), ha='center')
-    ax.set_title(title)
-    ax.set_xlabel(xlabel)
-    ax.set_ylabel(ylabel)
 
 datasets = ['davis', 'kiba']
 
@@ -115,48 +144,60 @@ if __name__ == "__main__":
 
     top_10_drug_mae = drug_mae.tail(10)
     top_10_drug_mae = top_10_drug_mae['compound_iso_smiles'].tolist()
-    with open(f'predictions/annotations/{model_st}_{dataset}{mutation}_drugs.json', 'w') as f:
-        json.dump(top_10_drug_mae, f, indent=4)
+    if ((not os.path.isfile(f'predictions/annotations/{model_st}_{dataset}{mutation}_drugs.json'))):
+        with open(f'predictions/annotations/{model_st}_{dataset}{mutation}_drugs.json', 'w') as f:
+            json.dump(top_10_drug_mae, f, indent=4)
 
     top_10_prot_mae = protein_mae.tail(10)
     top_10_prot_mae = top_10_prot_mae['target_sequence'].tolist()
-    with open(f'predictions/annotations/{model_st}_{dataset}{mutation}_proteins.json', 'w') as f:
-        json.dump(top_10_prot_mae, f, indent=4)
+    if ((not os.path.isfile(f'predictions/annotations/{model_st}_{dataset}{mutation}_proteins.json'))):
+        with open(f'predictions/annotations/{model_st}_{dataset}{mutation}_proteins.json', 'w') as f:
+            json.dump(top_10_prot_mae, f, indent=4)
+
+    if mutation != '':
+        proteins = 'data/{dataset}/new_proteins.json'
+    else:
+        proteins = f'data/{dataset}/proteins.txt'
+    
+    drugs = f'data/{dataset}/ligands_can.txt'
+
+    annotations_drug = f'predictions/annotations/{model_st}_{dataset}{mutation}_drugs.json'
+    annotations_prot = f'predictions/annotations/{model_st}_{dataset}{mutation}_proteins.json'
+
+    matching_keys_drugs = find_matching_keys(annotations_drug, drugs, canonicalize=True)
+    print(matching_keys_drugs.keys())
+    drug_keys = list(matching_keys_drugs.keys())
+
+    matching_keys_prots = find_matching_keys(annotations_prot, proteins)
+    print(matching_keys_prots.keys())
+    prot_keys = list(matching_keys_prots.keys())
 
     # Create the plot
-    fig, axs = plt.subplots(1, 2, figsize=(14, 6))
+    fig, axs = (plt.subplots(1, 2, figsize=(14, 6)))
 
-    # # Plot for drug_mae_sorted
-    # highlight_drugs = drug_mae.head(10)['compound_iso_smiles'].tolist()  # Highlight top 10 drugs with highest MAE
-    # plot_mae(drug_mae, 'compound_iso_smiles', 'mae', 'Prediction Error for Drugs', 'Drug', 'Median of Absolute Errors', highlight_drugs, axs[0])
+    # Plot for Drug MAE
+    axs[0].scatter(range(len(drug_mae)), drug_mae['mae'], s=10)
+    for i in range(-10, 0):  # Annotate only the top 10 highest MAE
+        axs[0].annotate(drug_keys[-i - 1], (len(drug_mae) + i, drug_mae['mae'].iloc[i]), fontsize=8,
+                        xytext=(-10, 0), textcoords = 'offset points', ha='right')
+    axs[0].set_xlabel('Drug')
+    axs[0].set_ylabel('Median of Absolute Errors for Affinity Prediction')
+    # axs[0].set_title(f'{model_st} Prediction Error for {dataset}{mutation} Test Data')
+    axs[0].set_xticks([])
 
-    # # Plot for protein_mae_sorted
-    # highlight_proteins = protein_mae.head(10)['target_sequence'].tolist()  # Highlight top 10 proteins with highest MAE
-    # plot_mae(protein_mae, 'target_sequence', 'mae', 'Prediction Error for Proteins', 'Protein', 'Median of Absolute Errors', highlight_proteins, axs[1])
-
-    # # Plot for Drug MAE
-    # axs[0].scatter(range(len(drug_mae)), drug_mae['mae'], s=10)
-    # # for i in range(-10, 0):  # Annotate only the top 10 highest MAE
-    # #     axs[0].annotate(drug_mae['compound_iso_smiles'].iloc[i], (len(drug_mae) + i, drug_mae['mae'].iloc[i]), fontsize=8,
-    # #                     xytext=(-10, 0), textcoords = 'offset points', ha='right')
-    # axs[0].set_xlabel('Drug')
-    # axs[0].set_ylabel('Median of Absolute Errors for Affinity Prediction')
-    # # axs[0].set_title(f'{model_st} Prediction Error for {dataset}{mutation} Test Data')
-    # axs[0].set_xticks([])
-
-    # # Plot for Protein MAE
-    # axs[1].scatter(range(len(protein_mae)), protein_mae['mae'], s=10)
-    # # for i in range(-10, 0):  # Annotate only the top 10 highest MAE
-    # #     axs[1].annotate(protein_mae['target_sequence'].iloc[i], (len(protein_mae) + i, protein_mae['mae'].iloc[i]), fontsize=8,
-    # #                     xytext=(-10, 0), textcoords = 'offset points', ha='right')
-    # axs[1].set_xlabel('Protein')
-    # # axs[1].set_ylabel('Median of Absolute Errors for Affinity Prediction')
-    # # axs[1].set_title(f'{model_st} Prediction Error for {dataset}{mutation} Test Data')
-    # axs[1].set_xticks([])
+    # Plot for Protein MAE
+    axs[1].scatter(range(len(protein_mae)), protein_mae['mae'], s=10)
+    for i in range(-10, 0):  # Annotate only the top 10 highest MAE
+        axs[1].annotate(prot_keys[-i - 1], (len(protein_mae) + i, protein_mae['mae'].iloc[i]), fontsize=8,
+                        xytext=(-10, 0), textcoords = 'offset points', ha='right')
+    axs[1].set_xlabel('Protein')
+    # axs[1].set_ylabel('Median of Absolute Errors for Affinity Prediction')
+    # axs[1].set_title(f'{model_st} Prediction Error for {dataset}{mutation} Test Data')
+    axs[1].set_xticks([])
     
-    # plt.suptitle(f'{model_st} Prediction Error for {dataset}{mutation} Test Data')
-    # plt.tight_layout()
-    # plt.show()
+    plt.suptitle(f'{model_st} Prediction Error for {dataset}{mutation} Test Data')
+    plt.tight_layout()
+    plt.show()
 
 # TO-DO:
 # load model - done
